fixed missing normalize flag usages and added withMutations to the schemas guide
