+++ docs/guides/schemas.md
@@ -112,3 +112,67 @@ This validation defines a very specific (honestly, useless) behavior, where if a
 When you need this level of specificity, using the `validateNode` property of the editor or plugins is handy.
 
 However, only use it when you absolutely have to. And when you do, you need to be aware of its performance. `validateNode` will be called **every time the node changes**, so it should be as performant as possible. That's why the example above returns early, so that the smallest amount of work is done as possible each time it is called.
+
+## Multi-step Normalizations
+
+Some normalizations will require multiple `change` function calls in order to complete. 
+
+Consider the following validation function that merges adjacent text nodes together. 
+
+Note: This functionality is already correctly implemented in slate-core so you don't need to put it in yourself!
+
+```
+/**
+  * Merge adjacent text nodes.
+  *
+  * @type {Object}
+  */
+validateNode(node) {
+  if (node.object != 'block' && node.object != 'inline') return
+
+  const invalids = node.nodes
+    .map((child, i) => {
+      const next = node.nodes.get(i + 1)
+      if (child.object != 'text') return
+      if (!next || next.object != 'text') return
+      return next
+    })
+    .filter(Boolean)
+
+  if (!invalids.size) return
+
+  return (change) => {
+    // Reverse the list to handle consecutive merges, since the earlier nodes
+    // will always exist after each merge.
+    invalids.reverse().forEach((n) => {
+      change.mergeNodeByKey(n.key)
+    })
+  }
+}
+```
+
+There is actually a problem with this code. Because each `change` function call will cause nodes impacted by the mutation to be normalized, this can cause interruptions to carefully implemented sequences of `change` functions and may create performance problems or errors. The normalization logic in the above example will merge the last node in the invalids list together, but then it'll trigger another normalization and start over!
+
+How can we deal with this? Well, normalization can be suppressed temporarily for multiple `change` function calls by using the `change.withMutations` function. `withMutations` accepts a function that takes a `change` object as a parameter, and executes the function while suppressing normalization. Once the function is done executing, the entire document is then normalized to pick up any unnnormalized transformations and ensure your document is in a normalized state.
+
+The above validation function can then be written as below
+
+```
+/**
+  * Merge adjacent text nodes.
+  *
+  * @type {Object}
+  */
+validateNode(node) {
+  ... 
+  return (change) => {
+    change.withMutations((c) => {
+      // Reverse the list to handle consecutive merges, since the earlier nodes
+      // will always exist after each merge.
+      invalids.reverse().forEach((n) => {
+        change.mergeNodeByKey(n.key)
+      })
+    });
+  }
+}
+```
