+++ packages/slate/src/changes/at-range.js
@@ -6,6 +6,7 @@ import Inline from '../models/inline'
 import Mark from '../models/mark'
 import Node from '../models/node'
 import String from '../utils/string'
+import getNormalizeSetting from '../utils/get-normalize-setting'
 
 /**
  * Changes.
@@ -28,7 +29,7 @@ const Changes = {}
 Changes.addMarkAtRange = (change, range, mark, options = {}) => {
   if (range.isCollapsed) return
 
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { value } = change
   const { document } = value
   const { startKey, startOffset, endKey, endOffset } = range
@@ -77,7 +78,7 @@ Changes.deleteAtRange = (change, range, options = {}) => {
   // when you undo a delete, the expanded selection will be retained.
   change.snapshotSelection()
 
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { value } = change
   let { startKey, startOffset, endKey, endOffset } = range
   let { document } = value
@@ -350,7 +351,7 @@ Changes.deleteWordBackwardAtRange = (change, range, options) => {
  */
 
 Changes.deleteBackwardAtRange = (change, range, n = 1, options = {}) => {
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { value } = change
   const { document } = value
   const { startKey, focusOffset } = range
@@ -536,7 +537,7 @@ Changes.deleteWordForwardAtRange = (change, range, options) => {
  */
 
 Changes.deleteForwardAtRange = (change, range, n = 1, options = {}) => {
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { value } = change
   const { document } = value
   const { startKey, focusOffset } = range
@@ -665,7 +666,7 @@ Changes.deleteForwardAtRange = (change, range, n = 1, options = {}) => {
 
 Changes.insertBlockAtRange = (change, range, block, options = {}) => {
   block = Block.create(block)
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
 
   if (range.isExpanded) {
     change.deleteAtRange(range)
@@ -717,7 +718,7 @@ Changes.insertBlockAtRange = (change, range, block, options = {}) => {
  */
 
 Changes.insertFragmentAtRange = (change, range, fragment, options = {}) => {
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
 
   // If the range is expanded, delete it first.
   if (range.isExpanded) {
@@ -830,7 +831,7 @@ Changes.insertFragmentAtRange = (change, range, fragment, options = {}) => {
  */
 
 Changes.insertInlineAtRange = (change, range, inline, options = {}) => {
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   inline = Inline.create(inline)
 
   if (range.isExpanded) {
@@ -906,7 +907,7 @@ Changes.insertTextAtRange = (change, range, text, marks, options = {}) => {
 Changes.removeMarkAtRange = (change, range, mark, options = {}) => {
   if (range.isCollapsed) return
 
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { value } = change
   const { document } = value
   const texts = document.getTextsAtRange(range)
@@ -936,7 +937,7 @@ Changes.removeMarkAtRange = (change, range, mark, options = {}) => {
  */
 
 Changes.setBlockAtRange = (change, range, properties, options = {}) => {
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { value } = change
   const { document } = value
   const blocks = document.getBlocksAtRange(range)
@@ -957,7 +958,7 @@ Changes.setBlockAtRange = (change, range, properties, options = {}) => {
  */
 
 Changes.setInlineAtRange = (change, range, properties, options = {}) => {
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { value } = change
   const { document } = value
   const inlines = document.getInlinesAtRange(range)
@@ -978,7 +979,7 @@ Changes.setInlineAtRange = (change, range, properties, options = {}) => {
  */
 
 Changes.splitBlockAtRange = (change, range, height = 1, options = {}) => {
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
 
   if (range.isExpanded) {
     change.deleteAtRange(range, { normalize })
@@ -1012,7 +1013,7 @@ Changes.splitBlockAtRange = (change, range, height = 1, options = {}) => {
  */
 
 Changes.splitInlineAtRange = (change, range, height = Infinity, options = {}) => {
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
 
   if (range.isExpanded) {
     change.deleteAtRange(range, { normalize })
@@ -1051,7 +1052,7 @@ Changes.toggleMarkAtRange = (change, range, mark, options = {}) => {
 
   mark = Mark.create(mark)
 
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { value } = change
   const { document } = value
   const marks = document.getActiveMarksAtRange(range)
@@ -1077,7 +1078,7 @@ Changes.toggleMarkAtRange = (change, range, mark, options = {}) => {
 Changes.unwrapBlockAtRange = (change, range, properties, options = {}) => {
   properties = Node.createProperties(properties)
 
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { value } = change
   let { document } = value
   const blocks = document.getBlocksAtRange(range)
@@ -1169,7 +1170,7 @@ Changes.unwrapBlockAtRange = (change, range, properties, options = {}) => {
 Changes.unwrapInlineAtRange = (change, range, properties, options = {}) => {
   properties = Node.createProperties(properties)
 
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { value } = change
   const { document } = value
   const texts = document.getTextsAtRange(range)
@@ -1216,7 +1217,7 @@ Changes.wrapBlockAtRange = (change, range, block, options = {}) => {
   block = Block.create(block)
   block = block.set('nodes', block.nodes.clear())
 
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { value } = change
   const { document } = value
 
@@ -1286,7 +1287,7 @@ Changes.wrapBlockAtRange = (change, range, block, options = {}) => {
 Changes.wrapInlineAtRange = (change, range, inline, options = {}) => {
   const { value } = change
   let { document } = value
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { startKey, startOffset, endKey, endOffset } = range
 
   if (range.isCollapsed) {
@@ -1395,7 +1396,7 @@ Changes.wrapInlineAtRange = (change, range, inline, options = {}) => {
  */
 
 Changes.wrapTextAtRange = (change, range, prefix, suffix = prefix, options = {}) => {
-  const { normalize = true } = options
+  const normalize = getNormalizeSetting(change.flags, options)
   const { startKey, endKey } = range
   const start = range.collapseToStart()
   let end = range.collapseToEnd()
