+++ packages/slate/src/changes/by-key.js
@@ -3,6 +3,7 @@ import Block from '../models/block'
 import Inline from '../models/inline'
 import Mark from '../models/mark'
 import Node from '../models/node'
+import getNormalizeSetting from '../utils/get-normalize-setting'
 
 /**
  * Changes.
@@ -11,26 +12,6 @@ import Node from '../models/node'
  */
 
 const Changes = {}
-
-/**
- * Obtains the normalization setting to use for the change based on the
- * change object flags and the change operation options.
- *
- * @param {Object} flags - flags object for the current change object
- * @param {Object} options - options for the change operation
- * @return {bool} - the normalization configuration to use. Defaults to true
- *   if flags or options does not have a normalization configuration
- */
-
-function getNormalizeSetting(flags, options) {
-  let normalize = true
-  if (flags.normalize !== undefined) {
-    normalize = flags.normalize
-  } else if (options.normalize !== undefined) {
-    normalize = options.normalize
-  }
-  return normalize
-}
 
 /**
  * Add mark to text at `offset` and `length` in node by `key`.
